// src/modules/pools/dto/wrap-pool.dto.ts
export class WrapPool {
  pairId: number;
  price: string;
  id: string;
  timestamp: number;
  tokenAmountIn: string;
  tokenIn: string;
  tokenOut: string;
  tokenAmountOut: string;
  tokenInSym: string;
  tokenOutSym: string;
  poolAddress: {
    id: string;
  };
  userAddress: {
    id: string;
  };
  poolLiquidity: string;
  isBuy: boolean;
  feeValue: string;
}